added tests with refactored calculator
