+++ tests/calculators/test_orca.py
@@ -1,89 +1,168 @@
-from os import path
-from shutil import rmtree
+"""
+This is testing the ORCA calculator given here, with an example orca output given to test on.
+
+the assets/ directory contains the orca files and this test depends on them
+"""
+
+import os
+import shutil
+
+import pytest
 
 import numpy as np
-from ase import Atoms
 from ase.build import molecule
-from ase.calculators.calculator import CalculationFailed
-from pytest import approx, raises
+from pytest import approx
 
-from wfl.calculators.orca.bh import BasinHoppingORCA
+from ase import Atoms
+from ase.calculators.calculator import CalculationFailed 
 
+from wfl.calculators.orca import ORCA
+from wfl.calculators import generic
+from wfl.configset import ConfigSet_in, ConfigSet_out
 
-def test_orca_utils(tmp_path):
-    # setup
-    at_ch4 = molecule("CH4")
-    at_ch3 = molecule("CH3")
+ref_parameters = dict(charge=0,
+                      mult=1,
+                      orca_command="dummy_no_orca_exec",
+                      orcablocks='%scf Convergence Tight \nSmearTemp 5000.0 \nmaxiter 500 \n  Rotate \n  {  1,   7,      28.37, 0, 0} \n  {  6,   9,      54.55, 0, 0} \n  {  2,  10,      27.42, 0, 0} \n  {  0,  14,      8.014, 1, 1} \n  {  3,  15,       35.3, 1, 1} \n  {  5,  13,      51.31, 1, 1} \n  {  2,  10,      17.84, 1, 1} \n  {  1,  11,      59.83, 1, 1} \nend \n \nend \n',
+                      orcasimpleinput='UHF revPBE def2-TZVP def2/J D3BJ slowconv',
+                      task='gradient')
 
-    calc_even = BasinHoppingORCA(scratch_path=tmp_path)
-    calc_even.atoms = at_ch4
-    calc_odd = BasinHoppingORCA()
-    calc_odd.atoms = at_ch3
-
-    # HOMO
-    assert calc_even.get_homo() == (4, 4)
-    assert calc_odd.get_homo() == (4, 3)
-
-    # multiplicity
-    assert calc_even.get_multiplicity() == 1
-    assert calc_odd.get_multiplicity() == 2
-
-    # new atoms
-    new_at_even: Atoms = calc_even._copy_atoms()
-    assert new_at_even.positions == approx(at_ch4.positions)
-    assert new_at_even.get_atomic_numbers() == approx(at_ch4.get_atomic_numbers())
-
-    # scratch paths
-    generated_scratch_dir = calc_even._make_tempdir()
-    assert f"{tmp_path}/orca_" in generated_scratch_dir
-
-    if path.isdir(generated_scratch_dir):
-        rmtree(generated_scratch_dir)
+ref_output = dict(energy=-2933.086022884649,
+                  forces=np.array([[-1.00418967e+03, -3.42059590e-07, 1.02844648e-05],
+                                   [1.00418967e+03, 3.42059590e-07, -1.02844648e-05]]),
+                  dipole=np.array([1.44071, -0.0, -0.0])
+                  )
 
 
-def test_orca_process_results(tmp_path):
-    # setup
-    at_ch4 = molecule("CH4")
-    calc = BasinHoppingORCA(scratch_path=tmp_path, forces_tol=0.05)
-    calc.atoms = at_ch4
+def test_extended_orca_calculator():
+    orca = ORCA()
+    orca.read(os.path.join(os.path.abspath(os.path.dirname(__file__)), '..', 'assets', 'orca'))
 
-    # shape errors, correct are (3, 10) and (3, 10, 5, 3)
-    with raises(AssertionError):
-        _ = calc.process_results(np.zeros(shape=(2, 10)), np.zeros(shape=(3, 10, 5, 3)))
-        _ = calc.process_results(np.zeros(shape=(3, 10)), np.zeros(shape=(3, 100, 5, 3)))
+    # parameter reading
+    for key, val in orca.parameters.items():
+        assert ref_parameters[key] == val
 
-    # none succeeding
-    e = np.zeros(shape=(3, 10)) + np.inf
-    f = np.zeros(shape=(3, 10, 5, 3))
-    with raises(CalculationFailed, match="Not enough runs succeeded.*0.*in wavefunction basin hopping"):
-        _ = calc.process_results(e, f)
+    # results reading
+    for key, val in orca.results.items():
+        # the dipole can bave a little floating point issue
+        # because it is written up to 5 digits only
+        assert ref_output[key] == approx(val, rel=(1e-4 if key == "dipole" else 1e-7))
 
-    # one of three
-    e[0, 1] = 0.999
-    with raises(CalculationFailed, match="Not enough runs succeeded.*1.*in wavefunction basin hopping"):
-        _ = calc.process_results(e, f)
 
-    # 2 of three passes
-    e[1, 2] = 1.0001
-    _ = calc.process_results(e, f)
 
-    # energy difference
-    e[2, 4] = 0.0
-    with raises(CalculationFailed, match="Too high energy difference found: *1.0.*"):
-        _ = calc.process_results(e, f)
+def test_orca_is_converged():
+    """function to check convergence from orca's output."""
 
-    # force error not triggered on non-minimum frame's force diff
-    e[2, 4] = np.inf
-    f[0, -1] = 100
-    _ = calc.process_results(e, f)
+    ref_path = os.path.join(os.path.abspath(os.path.dirname(__file__)), '..', 'assets')
+    orca = ORCA()
 
-    # error on forces
-    f[1, 2, 0, 0] = 0.051
-    with raises(CalculationFailed, match="Too high force difference found: *0.051.*"):
-        _ = calc.process_results(e, f)
+    orca.label = os.path.join(ref_path, 'orca_scf_converged')
+    assert orca.is_converged() is True
 
-    # results tested
-    f[0, 1] = 0.025
-    e_out, force_out = calc.process_results(e, f)
-    assert e_out == 0.999
-    assert force_out == approx(0.025)
+    orca.label = os.path.join(ref_path, 'orca_scf_unconverged')
+    assert orca.is_converged() is False
+
+    orca.label = os.path.join(ref_path, 'orca_scf_unfinished')
+    assert orca.is_converged() is None
+
+
+@pytest.mark.skipif(shutil.which("orca") is None, reason="no ORCA executable in path")
+def test_full_orca(tmp_path):
+    atoms = Atoms("H2", positions=[(0, 0, 0), (0, 0, 0.9)])
+
+    scratch_path = tmp_path / "fake_scratch_dir"
+    scratch_path.mkdir(exist_ok=True)
+    home_dir = tmp_path / "home_dir"
+
+    # this should raise error and copy over default files
+    calc = ORCA(directory=home_dir, 
+                scratch_path=scratch_path, 
+                keep_files = False, 
+                mult=2)
+
+    atoms.calc = calc
+    try:
+        atoms.get_potential_energy()
+    except CalculationFailed:
+        pass
+    assert not any(scratch_path.iterdir())
+    wdir_base = (home_dir / "ORCA_calc_files")
+    assert wdir_base.exists()
+    calc_dir = [d for d in wdir_base.iterdir()][0]
+    for ext in [".ase", ".inp", ".out"]:
+        assert (calc_dir / ("orca" + ext)).exists()
+
+    # just check this executes without error
+    calc = ORCA(directory=home_dir, 
+                scratch_path=scratch_path, 
+                keep_files = "default", 
+                mult=1)
+
+    atoms.calc = calc
+    atoms.get_potential_energy()
+
+
+@pytest.mark.skipif(shutil.which("orca") is None, reason="no ORCA executable in path")
+def test_orca_with_generic(tmp_path):
+    
+    home_dir = tmp_path / "home_dir"
+
+    atoms = Atoms("H2", positions=[(0, 0, 0), (0, 0, 0.9)])
+    atoms = [atoms] * 3 + [Atoms("H")]
+    inputs = ConfigSet_in(input_configs=atoms)
+    outputs = ConfigSet_out()
+
+    calc = ORCA(directory=home_dir, 
+                keep_files = "default", 
+                mult=1)
+
+    generic.run(inputs=inputs, outputs=outputs, calculator=calc, properties=["energy", "forces"], output_prefix="orca_", npool=0)
+    for at in outputs.to_ConfigSet_in():
+        assert "orca_energy" in at.info or "orca_calculation_failed" in at.info
+
+
+
+ref_freq = {'normal_mode_eigenvalues': np.array([0., 0., 0., 0., 0., 0., -0.60072079, -0.10155918, -0.07241669,
+                                                 -0.07238941, 0.14857026, 0.36614641, 0.37056397, 0.41033562,
+                                                 0.41034182]),
+            'normal_mode_displacements': np.array(
+                [[[0.0, 0.0, 0.0], [0.0, 0.0, 0.0], [0.0, 0.0, 0.0], [0.0, 0.0, 0.0], [0.0, 0.0, 0.0]],
+                 [[0.0, 0.0, 0.0], [0.0, 0.0, 0.0], [0.0, 0.0, 0.0], [0.0, 0.0, 0.0], [0.0, 0.0, 0.0]],
+                 [[0.0, 0.0, 0.0], [0.0, 0.0, 0.0], [0.0, 0.0, 0.0], [0.0, 0.0, 0.0], [0.0, 0.0, 0.0]],
+                 [[0.0, 0.0, 0.0], [0.0, 0.0, 0.0], [0.0, 0.0, 0.0], [0.0, 0.0, 0.0], [0.0, 0.0, 0.0]],
+                 [[0.0, 0.0, 0.0], [0.0, 0.0, 0.0], [0.0, 0.0, 0.0], [0.0, 0.0, 0.0], [0.0, 0.0, 0.0]],
+                 [[0.0, 0.0, 0.0], [0.0, 0.0, 0.0], [0.0, 0.0, 0.0], [0.0, 0.0, 0.0], [0.0, 0.0, 0.0]],
+                 [[-0.0, 0.0, 0.0], [0.0, -0.0, 4.99976e-01], [0.0, 0.0, 5.00019e-01], [0.0, -0.0, -5.01584e-01],
+                  [0.0, -0.0, -4.98416e-01]],
+                 [[0.0, 0.0, -1.65530e-01], [-0.0, 0.0, 4.93139e-01], [-0.0, 0.0, 4.93060e-01],
+                  [-0.0, -0.0, 4.91408e-01], [0.0, -0.0, 4.94797e-01]],
+                 [[-5.67750e-02, -8.91300e-02, -0.0], [3.78384e-01, -5.96550e-02, 0.0],
+                  [3.75511e-01, -5.71360e-02, 0.0], [-3.87060e-02, 5.90964e-01, 0.0], [-3.86700e-02, 5.87873e-01, 0.0]],
+                 [[-8.91340e-02, 5.67820e-02, -0.0], [5.88492e-01, 3.66800e-02, -0.0], [5.90352e-01, 4.06260e-02, -0.0],
+                  [-5.84790e-02, -3.77962e-01, 0.0], [-5.82730e-02, -3.75934e-01, 0.0]],
+                 [[8.00000e-06, -5.19000e-04, 0.0], [-5.00018e-01, 1.17000e-04, 0.0], [4.99974e-01, 1.72500e-03, 0.0],
+                  [7.89000e-04, -4.97828e-01, -0.0], [-8.44000e-04, 5.02166e-01, -0.0]],
+                 [[3.94000e-04, -3.90000e-05, 0.0], [-6.71000e-04, -4.99775e-01, -0.0],
+                  [-2.27900e-03, 5.00214e-01, -0.0], [-5.00876e-01, -8.01000e-04, 0.0],
+                  [4.99126e-01, 8.30000e-04, 0.0]],
+                 [[-8.30000e-05, -3.80000e-05, -0.0], [-6.92000e-04, -4.99738e-01, -0.0],
+                  [-2.27100e-03, 5.00261e-01, -0.0], [5.01967e-01, 7.96000e-04, 0.0],
+                  [-4.98019e-01, -8.67000e-04, 0.0]],
+                 [[6.77020e-02, -8.99660e-02, 0.0], [1.94210e-02, 5.61489e-01, 0.0], [1.60870e-02, 5.61639e-01, 0.0],
+                  [-4.19742e-01, -2.57180e-02, -0.0], [-4.22479e-01, -2.54060e-02, -0.0]],
+                 [[8.99710e-02, 6.77070e-02, 0.0], [2.42810e-02, -4.21873e-01, 0.0], [2.67800e-02, -4.20349e-01, 0.0],
+                  [-5.59691e-01, 1.79000e-02, -0.0], [-5.63437e-01, 1.75500e-02, -0.0]]])}
+
+
+@pytest.mark.skip(reason="Normal mode (eigenvector) reading implemented incorrectly.")
+def test_read_frequencies():
+    mol = molecule("CH4")
+    orca = ORCA(label=os.path.join(os.path.abspath(os.path.dirname(__file__)), '..', 'assets', 'orca_freq_dummy'))
+    orca.atoms = mol
+
+    # read the file given
+    orca.read_frequencies()
+
+    # results reading
+    for key, val in orca.extra_results.items():
+        assert ref_freq[key] == approx(val)
