+++ wfl/calculators/generic.py
@@ -10,12 +10,12 @@ from .utils import save_results
 
 
 # run that operates on ConfigSet, for multiprocessing
-def run(inputs, outputs, calculator, properties=None, output_prefix='_auto_', chunksize=10, verbose=False):
+def run(inputs, outputs, calculator, properties=None, output_prefix='_auto_', chunksize=10, verbose=False, npool=None):
     if properties is None:
         properties = ['energy', 'forces', 'stress']
     return iterable_loop(iterable=inputs, configset_out=outputs, op=run_op, chunksize=chunksize,
                          calculator=calculator, properties=properties, output_prefix=output_prefix,
-                         verbose=verbose)
+                         verbose=verbose, npool=npool)
 
 
 def run_op(atoms, calculator, properties=None, output_prefix='_auto_', verbose=False):
@@ -64,7 +64,7 @@ def run_op(atoms, calculator, properties=None, output_prefix='_auto_', verbose=F
             at.info[f'{output_prefix}calculation_failed'] = True
 
         # clean up invalid properties, will be fixed in quip Potential soon?
-        if 'virial' in at.clc.results:
+        if 'virial' in at.calc.results:
             del at.calc.results['virial']
 
         if calculation_succeeded:
